+++ lib/buffer/single/mmap_cache_test.go
@@ -30,6 +30,8 @@ import (
 )
 
 func TestMmapCacheTracker(t *testing.T) {
+	t.Skip("DEPRECATED")
+
 	dir, err := ioutil.TempDir("", "benthos_test_")
 	if err != nil {
 		t.Error(err)
@@ -118,6 +120,8 @@ func TestMmapCacheTracker(t *testing.T) {
 }
 
 func TestMmapCacheIndexes(t *testing.T) {
+	t.Skip("DEPRECATED")
+
 	dir, err := ioutil.TempDir("", "benthos_test_")
 	if err != nil {
 		t.Error(err)
@@ -216,6 +220,8 @@ func TestMmapCacheIndexes(t *testing.T) {
 }
 
 func TestMmapCacheRaces(t *testing.T) {
+	t.Skip("DEPRECATED")
+
 	dir, err := ioutil.TempDir("", "benthos_test_")
 	if err != nil {
 		t.Error(err)